refactor(vite): Update ui config to reuse code on cjs
