+++ packages/ui/config-cjs/helpers.cjs
@@ -1,5 +1,5 @@
 const colorToRGBA = require('./colorToRGBA.cjs')
-const PALLETE = require('./pallete.cjs')
+const PALLETE = require('./pallete.json')
 const generateColorsFromPalette = require('./generateColorsFromPalette.cjs')
 const u = require('./u.cjs')
