+++ packages/ui/config-cjs/rootConfig.cjs
@@ -1,2 +1,2 @@
-const config = require('../config.cjs')
+const config = require('../config-cjs')
 module.exports = config()
