+++ packages/ui/config/helpers.js
@@ -1,6 +1,1 @@
-import colorToRGBA from './colorToRGBA'
-import PALLETE from './pallete'
-import generateColorsFromPalette from './generateColorsFromPalette'
-import u from './u'
-
-export { PALLETE, generateColorsFromPalette, u, colorToRGBA }
+export { PALLETE, generateColorsFromPalette, u, colorToRGBA } from '../config-cjs/helpers.cjs'
