+++ packages/ui/config/rootConfig.js
@@ -1,2 +1,2 @@
-import config from '../config'
+import config from '../config-cjs'
 export default config()
